Fix and rename dataset script

+++ src/createdataset.py
@@ -8,17 +8,18 @@ from scapy.layers.tls import *
 
 INFILE = 'input.pcap'
 OUTDIR = 'dataset'
+PREFIX = 'pcap_output_'
 
 def make_dataset_tcprewrite():
     print('Creating tcprewrite dataset')
     #1 change ip address
-    subprocess.run(['tcprewrite', '--pnat=10.3.4.42:10.152.1.200', '--infile=./'+INFILE, '--outfile='+OUTDIR+'/out-01.pcap'])
+    subprocess.run(['tcprewrite', '--pnat=10.3.4.42:10.152.1.200', '--infile=./'+INFILE, '--outfile='+OUTDIR+'/'+PREFIX+'01.pcap'])
 
     #2 change ports
-    subprocess.run(['tcprewrite', '--portmap=80:8080,22:8022', '--fixcsum' , '--infile=./'+INFILE, '--outfile='+OUTDIR+'/out-02.pcap'])
+    subprocess.run(['tcprewrite', '--portmap=80:8080,22:8022', '--fixcsum' , '--infile=./'+INFILE, '--outfile='+OUTDIR+'/'+PREFIX+'02.pcap'])
 
     #3 change ip address and ports
-    subprocess.run(['tcprewrite', '--pnat=10.2.2.27:10.3.8.88', '--portmap=22:10222,443:60123', '--fixcsum' , '--infile=./'+INFILE, '--outfile='+OUTDIR+'/out-03.pcap'])
+    subprocess.run(['tcprewrite', '--pnat=10.2.2.27:10.3.8.88', '--portmap=22:10222,443:60123', '--fixcsum' , '--infile=./'+INFILE, '--outfile='+OUTDIR+'/'+PREFIX+'03.pcap'])
 
 def make_dataset_scapy(pkts):
     print('Creating scapy dataset')
@@ -31,7 +32,7 @@ def make_dataset_scapy(pkts):
             if pkt[DNS].qd:
                 pkt[DNS].qd.qname = 'www.example.com'
     
-    wrpcap(OUTDIR+'/out-04.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'04.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #5 change dns domain in dns query for specific address
@@ -41,7 +42,7 @@ def make_dataset_scapy(pkts):
                 if pkt[DNS].qd.qname == 'onecollector.cloudapp.aria.akadns.net':
                     pkt[DNS].qd.qname = 'example.com'
     
-    wrpcap(OUTDIR+'/out-05.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'05.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #6 change payload in http request
@@ -51,7 +52,7 @@ def make_dataset_scapy(pkts):
                 payload = pkt.lastlayer()
                 payload.original = re.sub(r'\s\d|\S*\.com', 'example.com', str(payload.original))
     
-    wrpcap(OUTDIR+'/out-06.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'06.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #7 change payload in request and change ip address
@@ -64,7 +65,7 @@ def make_dataset_scapy(pkts):
             if pkt[IP].src == '10.3.4.42':
                 pkt[IP].src = '1.1.1.1'
         
-    wrpcap(OUTDIR+'/out-07.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'07.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #8 change tls message
@@ -75,7 +76,7 @@ def make_dataset_scapy(pkts):
                     if pkt[TLS].type == 23:
                         pkt[TLS].msg = b'example.com'
 
-    wrpcap(OUTDIR+'/out-08.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'08.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #9 change source mac address only for source ip address 10.2.2.27
@@ -84,7 +85,7 @@ def make_dataset_scapy(pkts):
             if pkt[IP].src == '10.2.2.27':
                 pkt.src = '00:00:00:00:00:01'
         
-    wrpcap(OUTDIR+'/out-09.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'09.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #10 change dest mac address only for dest ip address 4.122.55.7
@@ -93,7 +94,7 @@ def make_dataset_scapy(pkts):
             if pkt[IP].dst == '4.122.55.7':
                 pkt.dst = '00:00:00:00:00:02'
             
-    wrpcap(OUTDIR+'/out-10.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'10.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #11 change protocol to udp for source ip address 10.2.2.27
@@ -102,7 +103,7 @@ def make_dataset_scapy(pkts):
             if pkt[IP].src == '10.2.2.27':
                 pkt[IP].proto = 17
 
-    wrpcap(OUTDIR+'/out-11.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'11.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #12 remove payload for source ip address 10.2.2.27
@@ -111,7 +112,7 @@ def make_dataset_scapy(pkts):
             if pkt[IP].src == '10.2.2.27':
                 pkt.remove_payload()
 
-    wrpcap(OUTDIR+'/out-12.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'12.pcap', pkts)
 
     #13 change dns answer in dns response
     for pkt in pkts:
@@ -121,7 +122,7 @@ def make_dataset_scapy(pkts):
                     if pkt[IP].src == '4.122.55.3':
                         pkt[DNS].an = b'1.1.1.1'
             
-    wrpcap(OUTDIR+'/out-13.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'13.pcap', pkts)
 
     #14 change tls message type
     for pkt in pkts:
@@ -131,7 +132,7 @@ def make_dataset_scapy(pkts):
                     if pkt[TLS].type == 23:
                         pkt[TLS].type = 22
 
-    wrpcap(OUTDIR+'/out-14.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'14.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #15 change tls message len
@@ -142,7 +143,7 @@ def make_dataset_scapy(pkts):
                     if pkt[TLS].type == 23:
                         pkt[TLS].len = 100
 
-    wrpcap(OUTDIR+'/out-15.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'15.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
 
@@ -163,7 +164,7 @@ def make_dataset_multi(pkts):
                 if pkt.haslayer(TCP):
                     pkt[TCP].sport = 10223
 
-    wrpcap(OUTDIR+'/out-16.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'16.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #17 change protocol to udp
@@ -173,7 +174,7 @@ def make_dataset_multi(pkts):
                 pkt[IP].src = '1.1.1.1'
                 pkt[IP].proto = 17
         
-    wrpcap(OUTDIR+'/out-17.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'17.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #18 change protocol to icmp if it is dhcp, and change its source mac address and ip address
@@ -184,7 +185,7 @@ def make_dataset_multi(pkts):
                 pkt.src = '00:00:00:00:00:01'
                 pkt[IP].src = '1.1.1.1'
     
-    wrpcap(OUTDIR+'/out-18.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'18.pcap', pkts)
 
 def make_dataset_scapy_ext(pkts):
     print('Creating scapy_ext dataset')
@@ -201,7 +202,7 @@ def make_dataset_scapy_ext(pkts):
                 pkt[IP].src = '4.122.55.199'
                 pkt[IP].dst = '4.122.55.198'
     
-    wrpcap(OUTDIR+'/out-19.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'19.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
     #20 change IP address so that it comes before DNS answer
@@ -210,28 +211,47 @@ def make_dataset_scapy_ext(pkts):
             if pkt[IP].dst == '157.55.134.142':
                 pkt[IP].dst = '104.96.141.107'
 
-    wrpcap(OUTDIR+'/out-20.pcap', pkts)
+    wrpcap(OUTDIR+'/'+PREFIX+'20.pcap', pkts)
     pkts = copy.deepcopy(orig)
 
 
 
 def ensure_dir():
-    # make sure the output directory exists
-    print('Creating directory: '+OUTDIR)
-    subprocess.run(['mkdir -p '+OUTDIR], shell=True)
-    # remove all files in the output directory
-    subprocess.run(['rm -f '+OUTDIR+'/*'], shell=True)
+    '''
+    Check if output directory exists, if not create it
+    Args:
+
+    Returns:
+    '''
+    if os.path.isdir(OUTDIR):
+        print('Output directory already exists: '+OUTDIR)
+        print('Purging all files in output directory')
+        subprocess.run(['rm -f '+OUTDIR+'/*'], shell=True)
+    else:
+        print('Creating directory: '+OUTDIR)
+        subprocess.run(['mkdir -p '+OUTDIR], shell=True)
 
 def ensure_file():
-    # make sure INFILE exists
+    '''
+    Check if input file exists
+    Args:
+
+    Returns:
+    '''
     print('Checking for input file: '+INFILE)
     if not os.path.isfile(INFILE):
         print('Input file input.pcap does not exist in current directory')
         exit(1)
 
-def run_all(pkts):
-    make_dataset_tcprewrite()
+def run_all():
+    '''
+    Run all methods to create datasets
+    Args:
 
+    Returns:
+    '''
+    pkts = rdpcap(INFILE)
+    make_dataset_tcprewrite()
     orig = copy.deepcopy(pkts)
     make_dataset_scapy(pkts)
     pkts = copy.deepcopy(orig)
@@ -240,9 +260,7 @@ def run_all(pkts):
     make_dataset_scapy_ext(pkts)
 
 if __name__ == '__main__':
-    
     ensure_file()
-
+    ensure_dir()
     load_layer('tls')
-    pkts = rdpcap(INFILE)
-    run_all(pkts)
+    run_all()
